+++ CHANGES.md
@@ -1,6 +1,27 @@
 # CHANGELOG
 
-## 0.7.0 (???)
+## 0.7.3 (???)
+
+### Bug fixes
+
+-   Fix error in `write_dataframe` if input has a date column and
+    non-consecutive index values (#325).
+
+## 0.7.2 (2023-10-30)
+
+### Bug fixes
+
+-   Add `packaging` as a dependency (#320).
+-   Fix conversion of WKB to geometries with missing values when using
+    `pandas.ArrowDtype` (#321).
+
+## 0.7.1 (2023-10-26)
+
+### Bug fixes
+
+-   Fix unspecified dependency on `packaging` (#318).
+
+## 0.7.0 (2023-10-25)
 
 ### Improvements